requirements.txt
absl-py==2.3.1
accelerate==1.10.0
bert-score==0.3.13
bitsandbytes==0.47.0
datasets==3.6.0
evaluate==0.4.5
hf-xet==1.1.8
huggingface-hub==0.34.4
python-dotenv==1.1.1
matplotlib==3.10.5
numpy==2.1.2
pandas==2.3.1
pillow==11.0.0
python-dotenv==1.1.1
requests==2.32.5
rouge-score==0.1.2
sacrebleu==2.5.1
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.1
tokenizers==0.21.4
torch==2.8.0+cu126
torchvision==0.23.0+cu126
tqdm==4.67.1
transformers==4.55.2